requirements.txt
discord.py>=2.5.0
python-dotenv>=1.1.0
motor>=3.7.1